renamed file with breadcrumbs content
