+++ openpype/tools/settings/settings/breadcrumbs_widget.py
